[#80] PartyCameraView 충돌 수정
