+++ MC2-OneShot/Presentation/View/CameraPreviewView.swift
@@ -1,0 +1,39 @@
+//
+//  CameraPreviewView.swift
+//  MC2-OneShot
+//
+//  Created by 김민준 on 5/21/24.
+//
+
+import SwiftUI
+import AVFoundation
+
+struct CameraPreviewView: UIViewRepresentable {
+    class VideoPreviewView: UIView {
+        override class var layerClass: AnyClass {
+            AVCaptureVideoPreviewLayer.self
+        }
+        
+        var videoPreviewLayer: AVCaptureVideoPreviewLayer {
+            return layer as! AVCaptureVideoPreviewLayer
+        }
+    }
+    
+    let session: AVCaptureSession
+    
+    func makeUIView(context: Context) -> VideoPreviewView {
+        let view = VideoPreviewView()
+        
+        view.videoPreviewLayer.session = session
+        view.backgroundColor = .black
+        view.videoPreviewLayer.videoGravity = .resizeAspectFill
+        view.videoPreviewLayer.cornerRadius = 0
+        view.videoPreviewLayer.connection?.videoOrientation = .portrait // TODO: 나중에 수정할 것!
+        
+        return view
+    }
+    
+    func updateUIView(_ uiView: VideoPreviewView, context: Context) {
+        
+    }
+}
